Flatten kd tree outside of constructor.
